Avoid race condition in detach/attach/delete instance
since detach/attach volume and delete instance are not synchronized
in Nova side, sometimes, volume will be left in a dilemma that described
in https://bugs.launchpad.net/cinder/+bug/1335889.

This patch is to avoid instance deletion while detach/attach volumes.
some basic test is performaned:
Test1:
1. create a instance
2. attach a volume and delete the instance in the same time
volume will be in 'available' status instead of previous 'in-use'
Test2:
1. create a instance
2. attach a volume to instance
2. detach a volume and delete the instance
The expect result is same as Test1